Create Task 1B - Naïve Bayesian Classifier for Text Classification
Description:
This task involves implementing a Naïve Bayesian classifier to perform text classification. The classifier will be trained on a set of documents and used to classify new documents. The implementation can utilize built-in Java classes/APIs. Accuracy, precision, and recall will be calculated to evaluate the performance of the classifier.
Explanation:

Training: The classifier is trained on a set of documents and their corresponding labels by counting the frequency of each word in each class.
Prediction: The class of a new document is predicted based on the calculated probabilities using the Naïve Bayesian formula.
Evaluation: Accuracy, precision, and recall are calculated to evaluate the classifier's performance on a test set.